darwin: abandon automated default settings and roll post-install into general setup
Working out defaults is too time consuming for the value provided. Various
settings are spread out across different namespaces and matching preferences set
to the values in the plists does not necessarily discover all necessary settings
required for a preference to be actually set. It is also very problematic when
moving across OSX versions where everything has to be re-verified and settings
re-discovered.